backend/requirements.txt
fastapi[standard]>=0.115
uvicorn[standard]>=0.30
sqlalchemy[asyncio]>=2.0
asyncpg>=0.30
alembic>=1.14
pydantic-settings>=2.6
python-multipart>=0.0.15
openai>=1.50
langgraph>=0.2
langchain-core>=0.3
pymupdf>=1.25
python-docx>=1.1
phonenumbers>=8.13
httpx>=0.28
pgvector>=0.3
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
